+++ go/lib/pktdisp/disp.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package disp
+package pktdisp
 
 import (
 	log "github.com/inconshreveable/log15"
